+++ CHANGELOG.md
@@ -7,11 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.1] - 30-10-20
-### Changed
-- update microkelvin  to `v0.5`
-- removal of unused canonical_host
+## [0.8.0] - 2021-04-21
 
-## [0.6.0] - 26-10-20
 ### Changed
-- update microkelvin and canonical to `v0.4`
+- Change microkelvin and canonical versions to 0.6 and 0.7 respectively
+
+## [0.7.0] - 2021-01-25
+
+### Changed
+- Change microkelvin and canonical versions to 0.5 and 0.6 respectively
+
+## [0.6.1] - 2020-10-30
+
+### Changed
+- Change microkelvin version to `v0.5`
+
+### Removed
+- Remove unused canonical_host
+
+## [0.6.0] - 2020-10-20
+### Changed
+- Change microkelvin and canonical to version 0.4
+
+[Unreleased]: https://github.com/dusk-network/nstack/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/dusk-network/nstack/compare/v0.7.0...v0.8.0
+[0.7.0]: https://github.com/dusk-network/nstack/compare/v0.6.1...v0.7.0
+[0.6.1]: https://github.com/dusk-network/nstack/compare/v0.6.0...v0.6.1
+[0.6.0]: https://github.com/dusk-network/nstack/releases/tag/v0.6.0